requirements.txt
django
djangorestframework
python-dotenv
google-generativeai